research-pipeline/requirements.txt
anthropic
python-dotenv
discord-webhook
